KeccakFamily parent struct
