+++ benches/tiny-keccak.rs
@@ -27,10 +27,34 @@ fn bench_sha3_256_input_4096_bytes(b: &mut Bencher) {
 #[bench]
 fn keccakf_u64(b: &mut Bencher) {
     const BYTES: usize = 25;
-    b.bytes = BYTES as u64;
+    b.bytes = (BYTES * 8) as u64;
 
     b.iter(|| {
         let mut data = [0u64; BYTES];
         keccakf(&mut data);
+    });
+}
+
+#[bench]
+fn bench_keccak256(b: &mut Bencher) {
+    const BYTES: usize = 32;
+    b.bytes = BYTES as u64;
+
+    b.iter(|| {
+        let data = [0u8; BYTES];
+        let mut result = [0u8; BYTES];
+        Keccak::keccak256(&data, &mut result);
+    });
+}
+
+#[bench]
+fn bench_k12(b: &mut Bencher) {
+    const BYTES: usize = 32;
+    b.bytes = BYTES as u64;
+
+    b.iter(|| {
+        let data = [0u8; BYTES];
+        let mut result = [0u8; BYTES];
+        k12(&[], &data, &mut result);
     });
 }
