+++ src/lib.rs
@@ -32,9 +32,10 @@
 #![no_std]
 
 mod kangaroo;
+mod keccak;
 
-use crunchy::unroll;
-pub use kangaroo::KangarooTwelve;
+pub use kangaroo::{k12, KangarooTwelve, keccakf as keccakf12};
+pub use keccak::*;
 
 const RHO: [u32; 24] = [
     1, 3, 6, 10, 15, 21, 28, 36, 45, 55, 2, 14, 27, 41, 56, 8, 25, 43, 62, 18, 39, 61, 20, 44,
@@ -44,98 +45,11 @@ const PI: [usize; 24] = [
     10, 7, 11, 17, 18, 3, 5, 16, 8, 21, 24, 4, 15, 23, 19, 13, 12, 2, 20, 14, 22, 9, 6, 1,
 ];
 
-const RC: [u64; 24] = [
-    1u64,
-    0x8082u64,
-    0x800000000000808au64,
-    0x8000000080008000u64,
-    0x808bu64,
-    0x80000001u64,
-    0x8000000080008081u64,
-    0x8000000000008009u64,
-    0x8au64,
-    0x88u64,
-    0x80008009u64,
-    0x8000000au64,
-    0x8000808bu64,
-    0x800000000000008bu64,
-    0x8000000000008089u64,
-    0x8000000000008003u64,
-    0x8000000000008002u64,
-    0x8000000000000080u64,
-    0x800au64,
-    0x800000008000000au64,
-    0x8000000080008081u64,
-    0x8000000000008080u64,
-    0x80000001u64,
-    0x8000000080008008u64,
-];
-
-/// keccak-f[1600]
-#[allow(unused_assignments)]
-pub fn keccakf(a: &mut [u64; PLEN]) {
-    for i in 0..24 {
-        let mut array: [u64; 5] = [0; 5];
-
-        // Theta
-        unroll! {
-            for x in 0..5 {
-                unroll! {
-                    for y_count in 0..5 {
-                        let y = y_count * 5;
-                        array[x] ^= a[x + y];
-                    }
-                }
-            }
-        }
-
-        unroll! {
-            for x in 0..5 {
-                unroll! {
-                    for y_count in 0..5 {
-                        let y = y_count * 5;
-                        a[y + x] ^= array[(x + 4) % 5] ^ array[(x + 1) % 5].rotate_left(1);
-                    }
-                }
-            }
-        }
-
-        // Rho and pi
-        let mut last = a[1];
-        unroll! {
-            for x in 0..24 {
-                array[0] = a[PI[x]];
-                a[PI[x]] = last.rotate_left(RHO[x]);
-                last = array[0];
-            }
-        }
-
-        // Chi
-        unroll! {
-            for y_step in 0..5 {
-                let y = y_step * 5;
-
-                unroll! {
-                    for x in 0..5 {
-                        array[x] = a[y + x];
-                    }
-                }
-
-                unroll! {
-                    for x in 0..5 {
-                        a[y + x] = array[x] ^ ((!array[(x + 1) % 5]) & (array[(x + 2) % 5]));
-                    }
-                }
-            }
-        };
-
-        // Iota
-        a[0] ^= RC[i];
-    }
-}
-
-/// Total number of lanes.
 const PLEN: usize = 25;
+
+trait Permutation {
+    fn execute(a: &mut [u64; PLEN]);
+}
 
 #[derive(Default, Clone)]
 struct Buffer([u64; PLEN]);
@@ -195,143 +109,43 @@ impl Buffer {
     }
 }
 
-#[derive(Debug, Clone, Copy)]
-enum Rounds {
-    Normal,
-    Reduced,
-}
-
-/// This structure should be used to create keccak/sha3 hash.
-///
-/// ```rust
-/// use tiny_keccak::Keccak;
-///
-/// fn main() {
-///     let mut sha3 = Keccak::new_sha3_256();
-///     let data: Vec<u8> = From::from("hello");
-///     let data2: Vec<u8> = From::from("world");
-///
-///     sha3.update(&data);
-///     sha3.update(&[b' ']);
-///     sha3.update(&data2);
-///
-///     let mut res: [u8; 32] = [0; 32];
-///     sha3.finalize(&mut res);
-///
-///     let expected = vec![
-///         0x64, 0x4b, 0xcc, 0x7e, 0x56, 0x43, 0x73, 0x04,
-///         0x09, 0x99, 0xaa, 0xc8, 0x9e, 0x76, 0x22, 0xf3,
-///         0xca, 0x71, 0xfb, 0xa1, 0xd9, 0x72, 0xfd, 0x94,
-///         0xa3, 0x1c, 0x3b, 0xfb, 0xf2, 0x4e, 0x39, 0x38
-///     ];
-///
-///     let ref_ex: &[u8] = &expected;
-///     assert_eq!(&res, ref_ex);
-/// }
-/// ```
-pub struct Keccak {
+struct KeccakFamily<P> {
     buffer: Buffer,
     offset: usize,
     rate: usize,
     delim: u8,
-    rounds: Rounds,
+    permutation: ::core::marker::PhantomData<P>,
 }
 
-impl Clone for Keccak {
+impl <P> Clone for KeccakFamily<P> {
     fn clone(&self) -> Self {
-        let mut res = Keccak::new_with_rounds(self.rate, self.delim, self.rounds);
-        res.buffer = self.buffer.clone();
-        res.offset = self.offset;
-        res
+        KeccakFamily {
+            buffer: self.buffer.clone(),
+            offset: self.offset,
+            rate: self.rate,
+            delim: self.delim,
+            permutation: ::core::marker::PhantomData,
+        }
     }
 }
 
-macro_rules! impl_constructor {
-    ($name: ident, $alias: ident, $bits: expr, $delim: expr) => {
-        pub fn $name() -> Keccak {
-            Keccak::new(200 - $bits / 4, $delim)
-        }
-
-        pub fn $alias(data: &[u8], result: &mut [u8]) {
-            let mut keccak = Keccak::$name();
-            keccak.update(data);
-            keccak.finalize(result);
-        }
-    };
-}
-
-macro_rules! impl_global_alias {
-    ($alias: ident, $size: expr) => {
-        pub fn $alias(data: &[u8]) -> [u8; $size / 8] {
-            let mut result = [0u8; $size / 8];
-            Keccak::$alias(data, &mut result);
-            result
-        }
-    };
-}
-
-impl_global_alias!(shake128, 128);
-impl_global_alias!(shake256, 256);
-impl_global_alias!(keccak224, 224);
-impl_global_alias!(keccak256, 256);
-impl_global_alias!(keccak384, 384);
-impl_global_alias!(keccak512, 512);
-impl_global_alias!(sha3_224, 224);
-impl_global_alias!(sha3_256, 256);
-impl_global_alias!(sha3_384, 384);
-impl_global_alias!(sha3_512, 512);
-
-impl Keccak {
-    pub fn new(rate: usize, delim: u8) -> Keccak {
-        Keccak::new_with_rounds(rate, delim, Rounds::Normal)
-    }
-
-    fn new_with_rounds(rate: usize, delim: u8, rounds: Rounds) -> Keccak {
+impl <P: Permutation> KeccakFamily<P> {
+    fn new(rate: usize, delim: u8) -> Self {
         assert!(rate != 0, "rate cannot be equal 0");
-        Keccak {
+        KeccakFamily {
             buffer: Buffer::default(),
             offset: 0,
             rate,
             delim,
-            rounds,
+            permutation: ::core::marker::PhantomData,
         }
     }
 
-    impl_constructor!(new_shake128, shake128, 128, 0x1f);
-    impl_constructor!(new_shake256, shake256, 256, 0x1f);
-    impl_constructor!(new_keccak224, keccak224, 224, 0x01);
-    impl_constructor!(new_keccak256, keccak256, 256, 0x01);
-    impl_constructor!(new_keccak384, keccak384, 384, 0x01);
-    impl_constructor!(new_keccak512, keccak512, 512, 0x01);
-    impl_constructor!(new_sha3_224, sha3_224, 224, 0x06);
-    impl_constructor!(new_sha3_256, sha3_256, 256, 0x06);
-    impl_constructor!(new_sha3_384, sha3_384, 384, 0x06);
-    impl_constructor!(new_sha3_512, sha3_512, 512, 0x06);
-
-    pub fn update(&mut self, input: &[u8]) {
-        self.absorb(input);
+    fn keccakf(&mut self) {
+        P::execute(self.buffer.inner())
     }
 
-    #[inline]
-    pub fn keccakf(&mut self) {
-        match self.rounds {
-            Rounds::Normal => keccakf(self.buffer.inner()),
-            Rounds::Reduced => kangaroo::keccakf(self.buffer.inner()),
-        }
-    }
-
-    pub fn finalize(mut self, output: &mut [u8]) {
-        self.pad();
-
-        // apply keccakf
-        self.keccakf();
-
-        // squeeze output
-        self.squeeze(output);
-    }
-
-    // Absorb input
-    pub fn absorb(&mut self, input: &[u8]) {
+    fn update(&mut self, input: &[u8]) {
         //first foldp
         let mut ip = 0;
         let mut l = input.len();
@@ -351,17 +165,11 @@ impl Keccak {
         self.offset = offset + l;
     }
 
-    pub fn pad(&mut self) {
+    fn pad(&mut self) {
         self.buffer.pad(self.offset, self.delim, self.rate);
     }
 
-    pub fn fill_block(&mut self) {
-        self.keccakf();
-        self.offset = 0;
-    }
-
-    // squeeze output
-    pub fn squeeze(&mut self, output: &mut [u8]) {
+    fn squeeze(&mut self, output: &mut [u8]) {
         // second foldp
         let mut op = 0;
         let mut l = output.len();
@@ -375,42 +183,13 @@ impl Keccak {
         self.buffer.setout(&mut output[op..], 0, l);
     }
 
-    #[inline]
-    pub fn xof(mut self) -> XofReader {
+    fn finalize(mut self, output: &mut [u8]) {
         self.pad();
 
+        // apply keccakf
         self.keccakf();
 
-        XofReader {
-            keccak: self,
-            offset: 0,
-        }
+        // squeeze output
+        self.squeeze(output);
     }
 }
-
-pub struct XofReader {
-    keccak: Keccak,
-    offset: usize,
-}
-
-impl XofReader {
-    pub fn squeeze(&mut self, output: &mut [u8]) {
-        // second foldp
-        let mut op = 0;
-        let mut l = output.len();
-        let mut rate = self.keccak.rate - self.offset;
-        let mut offset = self.offset;
-        while l >= rate {
-            self.keccak.buffer.setout(&mut output[op..], offset, rate);
-            self.keccak.keccakf();
-            op += rate;
-            l -= rate;
-            rate = self.keccak.rate;
-            offset = 0;
-        }
-
-        self.keccak.buffer.setout(&mut output[op..], offset, l);
-        self.offset = offset + l;
-    }
-}
-
